+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to `homebridge` will be documented in this file. This projec
 - fix(matter): add missing occupancy cluster to motion sensor
 - fix(matter): add missing switch cluster to generic switch
 - fix(matter): add missing smoke/co cluster to smoke sensor
+- fix(matter): add missing pump control cluster to pump
 
 ### Homebridge Dependencies